Shadow volume attached to Light in constructor as a test. Renders as if it were part of the light's radius.

+++ source/StaticGeometry.h
@@ -2,6 +2,7 @@
 #define STATIC_GEOMETRY_H
 
 #include "GL/gl.h"
+#include <stdint.h>
 #include <vector>
 
 using std::vector;
@@ -10,7 +11,7 @@ class StaticGeometry
 {
 private:
     //zones, portals, etc.
-    GLuint          zone;
+    int16_t          zone;
 
 public:
     void Render() const;
